requirements.txt
customtkinter>=5.2.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
pyinstaller>=6.0.0
